ヒーラルキーウィンドウ内のキャンバス関係の整理

+++ ZemiPhoton/Assets/Resources/Hasegawa/Script/Player/PlayerInfo.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class PlayerInfo : MonoBehaviour {
 
 	public static string playerName = "default";
-	public static byte playerNumber = 1;
+	public static int playerNumber = 1;
 
 	public static bool Spawn = false;
